Added remote B3 ability.
In b3.xml, simply set the game_log path in the syntax:
ftp://user:pass@host/path:logfilename, b3 will automatically download the file to your hard disk.

Basic functionality, if connection is lost, it should automatically restart, but it has not been tested, and if it was down for ten or more minutes on a populated server, there would certainly be severe problems. Expect a restart implementation within around 48 hours.

Signed-off-by: James 'Bakes' Baker <bakes@ymail.com>