[core] fix getStuffSoundingLike that would return non unique suggestions or would not return suggestions in the same order for the same input values if ordered differently and if having multiple same levenshtein distance.
What happened is that the `expected_stuff` was obtained from `some_dict.keys()` and the order of elements obtained by `keys()` is undefined and as such not guaranteed to be the same on Windows and Linux.
This highlighted the bug that occurs when multiple values from `expected_stuff` have the same levenshtein distance from `clean_stuff`; in such case we would return the `exepected_stuff` elements in the order they were found.